mgr/cephadm: Initialize with cluster version
When the VersionTracker class is initialized it will check if the sqlite
table is empty, if the table is empty the current cluster version will be
added, this will effectively ensure the initial clsuter version is added
to the table. Minor clarity updates were also added with fetchall and
variable name in sql query was fixed.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>