OSDMap::calc_pg_upmaps: accept more proposed changes
In some cases the balancer can stall out, unable to find any
changes that will help overfull or underfull OSDs get closer
to their target num PGs.

Here's an example from the wild. These changes are not accepted
because the overall stddev does not appear to be decreasing:

  will try dropping existing remapping pair 264 -> 137 which remapped 24.f9d into overfull osd.137
  existing pg_upmap_items [766,799,264,137] remapped 24.f9d into overfull osd.137, new_pg_upmap_items now [766,799]
  stddev 138085.453125 -> 138085.453125

  will try dropping existing remapping pair 496 -> 637 which remapped 24.5ca out from underfull osd.496
  existing pg_upmap_items [729,790,496,637] remapped 24.5ca out from underfull osd.496, new_pg_upmap_items now [729,790]
  stddev 138085.453125 -> 138085.453125

Those changes will now be accepted, which is good because those
OSDs will improve, even though the sum of squared deviations does
not.

Note that a change scoring exactly equal still moves data, so this
trades some churn for the ability to escape a plateau. The two
following patches make the score both more meaningful and precise
enough to tell many of these cases apart on their own merits.

Signed-off-by: Dan van der Ster <dan.vanderster@clyso.com>