make-debs.sh: invoke tar with --no-same-owner
When running as a normal user, tar does not attempt to preserve
owners set on the tar content files.  When running as root, it does.
Containerized builds are running as root.  Stop make-debs.sh from
trying to set other owners for files, and leaving files in the
host system with mapped UIDs other than the user running the container
(which causes jenkins to be unable to clear the workspace).

Signed-off-by: Dan Mick <dan.mick@redhat.com>