OSDMap::calc_pg_upmaps: score maps using weighted error
Currently we score how balanced we are using a sum of the absolute
deviations, treating small and large OSDs equally. This strongly
under values relatively small PGs, which we really need to be
accurately balancing with priority to avoid them getting overfull.

Consider this scenario:

  osd.0 weight 100.0 pgs 100 target 100 -> deviation 0
  osd.1 weight 1.0   pgs 2   target 1   -> deviation 1

The tiny OSD osd.1 is 2x overfull -- we should prioritize moving
one of its PGs to osd.0, which would only go 1% overfull. The
new values would become:

  osd.0 weight 100.0 pgs 101 target 100 -> deviation 1
  osd.1 weight 1.0   pgs 1   target 1   -> deviation 0

With the "stddev" calculation, we have these before/after scores:

  stddev = 0*0 + 1*1 = 1
  new_stddev = 1*1 + 0*0 = 1

(Until recently, that change would be rejected because it does not
decrease the stddev).

To improve things, we calculate a bal_score which incorporates the
relative deviation and OSD weights.

Here's the new scores for the example change above:

  bal_score = (100/100 - 1)^2 * 100 + (2/1 - 1)^2 * 1 = 2
  new_bal_score = (101/100 - 1)^2 * 100 + (1/1 - 1)^2 * 1 = 0.01

With this new approach the change is highly favourable, as it
should be.

Signed-off-by: Dan van der Ster <dan.vanderster@clyso.com>