src: replace push_back(T{}) with emplace_back()
in 1bfd396f, we switched from some instances of `push_back(Type{})`
patterns to `emplace_back()` to simplify the code and to avoid
unnecessary overhead of copy/move constructor calls. but there were
still remaining instances of them in `generate_test_instance()`
member function implementations.

in this change, we switched all of them to be more consistent.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>