qa/workunits/rbd: fix race in test_force_promote_delete_group
... that is a part of the rbd_mirror_group_simple test suite.

In "test_force_promote_delete_group," the secondary group is
force-promoted, but the original primary is not demoted. The
force-promoted group is disabled for mirroring. One of its member images
is removed. The group is then re-enabled for mirroring before being
removed.

Sometimes the check to confirm that the group has been removed fails.
This happens because the rbd-mirror daemon’s group replayer, after
restarting, recreates the group immediately after it has been explicitly
removed.

To fix this race, disable the original primary group. This prevents the
rbd-mirror daemon from recreating the force-promoted group that was
recently deleted.

Signed-off-by: Ramana Raja <rraja@redhat.com>