common: Add daemon uniquifier to PerfCounters
Dashboard has asked for a unique identifier that can be specified at
setup and returned in PerfCounters for RGW. Matt suggested we could do
something that works for all daemons in case another needs similar
functionality.

Thus, we provide the `--daemon_uniquifier` argument that can be passed
to any program using `global_init`. If non-zero, a `daemon` section is
added to the PerfCounters dump for that program with a `uniquifier`
entry set to the provided value. If unspecified or zero, no `daemon`
section is added.

The uniquifier may be any non-zero unsigned 64-bit integer.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>