cephfs.pyx: in rmtree(), don't run rmdir() to figure out if a dir...
is empty, instead add it to stack and break out of the inner loop. This
let's outer loop read this dir (since it is at top of the stack). If dir
is found empty, it is deleted, else, it is kept on stack and its dirs
are explored and its files are deleted.

Fixes: https://tracker.ceph.com/issues/72992
Signed-off-by: Rishabh Dave <ridave@redhat.com>
(cherry picked from commit 65dd5431d5df408a4fe44be1b3dfa6feaeda9ecf)