cephadm: added edge case checking for version tracking
In the case that someone upgrades to a Ceph version with version tracking but
previously didn't have a Ceph version with version tracking the bootstrap version
and bootstrap time are not recorded. To avoid errors I've added checks for if the
KV values are present so it doesn't try to pull non-existent values. I've also
added a add bootstrap cluster version function to be applied in get cluster version
history function and in the upgrade function to ensure the bootstrap version is
always present.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>