qa/workunits: add scenario to "test_force_promote_delete_group"
... in rbd_mirror_group_simple test suite.

After the group and its images are removed from the secondary, the test
can run in one of two scenarios. In Scenario 1, the test confirms that
the group is completely synced from the primary to the secondary. In
Scenario 2, the test disables and re-enables the primary, and then
confirms the group syncs from the primary to the secondary.

Scenario 2 occasionally fails. The disruption to the rbd-mirror daemon's
normal workflow from disabling and re-enabling the primary sometimes
prevents the mirror daemon from correctly syncing the group.

Signed-off-by: Ramana Raja <rraja@redhat.com>