mgr/cephadm: added check for mgr db in upgrade
In the upgrade process, there is now a check for if the mgr db is
ready, if it is not an error will raised to inform users of the issue.
This check is necessary because the mgr db is needed to store
cluster version history, in this case, the bootstrap version and
upgrade version.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>