cephadm: add uca prototype to cephadm subtree
What's uca?

uca is a tool that brings some of cephadm's capabilities to systems
that are not part of a ceph cluster. The goal is run certain commands
and automation without having to ssh into a ceph cluster (admin) node
first.

At present the command supports a shell subcommand that works similar
to `cephadm shell`. It also has a direct wrapper for `ceph` command line
tool similar to `cephadm shell ceph ...` but with a few additional
abilities.

What's uca mean?

Naming is hard. I had to call it something.
* uca could stand for "Userspace Ceph Admin"
* uca could stand for "micro ceph admin" (yeah yeah u ain't micro)
* uca could stand for "Uncanny Cephadm Approximate"

Choose one you like or suggest your own. The name is also not set in
stone yet anyway.

Why?

For a while (~2 years or so) I have wanted a tool like this. In my
previous experience working on k8s systems I always appreciated
that I could run `kubectl` or `oc` on my system to issue commands
and submit resources to the cluster. It's just a matter of having
the right keys and the ability to send packets to the API really.

At some point it dawned on me that the same is true of Ceph. Ceph
has cephx keys and APIs (rados). The tooling is a bit different,
a bit more "dispersed" but the Ceph project already packages things
in to containers and cephadm already knows how to execute those
containers in a way that allows you to manage the ceph cluster.

How's it work?

uca is basically a very stripped down version of cephadm. It shares
the libraries cephadm uses but it is limited to running commands
that make sense outside of a ceph node, so no systems management
stuff. But the containers, yeah, we can do that.

Unlike cephadm, uca is designed to work with multiple ceph clusters.
The `uca cluster ...` subcommands allow one to list, add, remove, etc
clusters. You can mark a cluster as default to make uca commands
use that cluster without additional arguments.

Do note that it requires "line of sight" to the Ceph cluster you
want to manage, as the tools within the container will need to
connect to rados/the mons in the ceph cluster.

Who can use it?

I designed it for my own use cases and have been using it on and off for
a couple of months now. Consider it developer-pre-preview.

Eventually, I see it useful for both developers as well as ceph cluster
admins. Imagine issuing commands from your laptop based on yaml
files stored in a git repo instead of having to copy them to an admin
node first.

How do I get started?

If you want to give it a whirl you need to build it manually at
the moment. Run:
```
./src/cephadm/build.py --main-script uca.py  /tmp/uca
```
and it will produce a python "zipapp" at /tmp/uca. Run
```
/tmp/uca -h
```
to see the help.

To initialize a cluster config from a host you can ssh into, run:
```
/tmp/uca cluster ssh-init --make-default  root@my.host.example.com
```
(see the help for other params)

You can also manually initialize a cluster using `uca cluster init`
and providing it with config and keyring files.

Now, as an example, run:
```
/tmp/uca ceph status
```
Or perhaps:
```
/tmp/uca ceph orch ls
```

You can enter the shell and use additional tools, like `rados` for
example:
```
$ /tmp/uca shell
[...]# rados lspools
cephfs.cephfs.meta
cephfs.cephfs.data
.mgr
.smb
```

Have fun!

Signed-off-by: John Mulligan <jmulligan@redhat.com>