qa/tasks: Ignore expected `backfill_toofull` warning
We deliberately inject an error to this test using
`osd_debug_reject_backfill_probability=0.3`, which
determines a >0 probability that a PG's backfill
reservation is rejected.

If the reservation is rejected, we will see this
type of error in the OSD log:
```
backfill reservation rejected: failure injection
```

Since this is an injected test scenario, we can
add this warning to the ignorelist.

Fixes: https://tracker.ceph.com/issues/70716
Signed-off-by: Laura Flores <lflores@ibm.com>