rbd-mirror: mirror_group_snapshot_unlink_peer should also care about local snaps
Problem:
In a relocate situation there is a demote involved and we have a demote
snapshot both on primary with peer UUID attached and on secondary with peer UUID
attached. Later a promote say on site-b triggered primary demote group snap
removal which started a image snap removal, so it started a request to
image replayer (idle at that time), but the request didn't yet went
through.

As a next command we had --force promote on site-a but since the request on
site-a for image snap removal is already in place it somehow went through and
removed the image snaps leaving the group snap on site-a primary demote snapshot
partially deleted with peer UUID attached. And now note site-a is again primary.

In the later test, it involved demote on site-b and followed by a resync on it.
But since the primary demote snapshot is lingering in the partially deleted
state and since peer UUID is still attached, there is no way the resync will
go through, leaving the syncing snaps in incomplete state forever.

Note: Any CLI command on new primary site-a cannot delete the partially deleted
snap because there is a peer UUID attached to it.

Solution:
Update mirror_group_snapshot_unlink_peer() to also inspect local snapshots
when unlinking peer of a remote group snapshot. Specifically:

  When removing a remote group snapshot:
    * If the remote snapshot is marked as non-primary,
    * Then check for a corresponding local snapshot with the peer UUID.
    * If such a local snapshot exists and:
      - It is in a primary state (i.e., it was created during a demote),
      - And it still holds a peer UUID,
      - Then unlink peer UUID from local snapshot as well to ensure consistency.

 This ensures the partially deleted demote snapshot does not linger with
 an peer UUID attached, thus:
   * Allowing future resyncs to succeed.
   * Preventing "partially deleted" snapshots from blocking mirroring logic.

Additionally this change fixes an old issue i.e. after relocate, on new primary
the previous primary demoted snap is not getting deleted, until the next
demotion. This case is handled now.

Signed-off-by: Prasanna Kumar Kalever <prasanna.kalever@redhat.com>