mgr/cephadm: added initial version tracking check in serve loop
In the CephadmServe main loop, I've added a line that checks if the mgr db is ready,
then it will call a function that adds the bootstrap version to a sqlite table that tracks
version history. This was added to ensure the initial version entry is added as early
as possible so that datetime value is more accurate for the entry. The function also
only adds the entry if the table is empty so there is no problem of repeated entry
additions.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>