rgw/obi: Initial set of mods supporting ordered bucket indexes
Throughout the code base bucket index shards are referenced by indexes
ranging from 0 to n-1 where n is the number of shards. That code needs
to start shifting.

The BIndexer class is created to contain all knowledge about mapping
objects to shards across indexing schemes. It's a virtual base class
and there are subclasses for each bucket index sharding strategy,
HashedBIndexer and OrderedBIndexer. All mapping should be routed
through a BIndexer class.

In general we'll use std::unique_ptr<rgw::rados::BIndexer>'s (or
possibly a shared_ptr) to get dynamic dispatch to the appropriate
subclass's functionality.

Within the bucket layout structures we'll use:

using LayoutVariant =
    std::variant<bucket_index_hashed_layout,bucket_index_ordered_layout>;

to store the various layout metadata at rest in the bucket instance
data structure.

And because we can no longer use for-loops an integral loop control
variable to visit each shard, BIndexer has a nested class of
ShardIterator, which in turn maintains an interior iterator
InteriorShardIterator that is implemented as a virtual base class.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>