osd: Check for read instability when doing EC direct reads
can_serve_replica_read(...) previously only applied to replicas as all reads in erasure coding would go through the primary. With the implementation of direct reads in EC, it is no longer the case that reads go through the primary, so we want to expand this logic to ec. The existing logic applies almost perfectly to EC in it's current state, so here we just need to update the function names and debug messages to reflect that this code is now used when going through the direct read path in ec as well.

Signed-off-by: Jon Bailey <jonathan.bailey1@ibm.com>