mgr/cephadm: added debug logs and sqlite error check
I added debug logs to all major adding cluster history operations,
these debug logs track whether a cluster version was successfully
added or not. The logs span several lines and exist in multiple
layered functions. They should be used to help developers find
out why a cluster may be missing cluster history and possibly
allow for the recovery of this history as the cluster version is
displayed in the debug logs. I have also added a general sqlite3
error check to all sqlite operations to ensure the version tracker
will not crash a cluster due to any unforseen sqlite3 errors. I have
also added the additional change that cluster upgrade is no longer
blocked if OSDs are not available and the SCHEMAs are logically
moved to version tracker for better organization.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>