crimson/rgw: Fix clang compiler warnings.
These fixes address compiler warnings that may only be present in
clang builds.

In RGW, the problem is a warning not implemented in clang. Despite
apparantly being a GCC-specific pragma, clang actually attempts to
honour the progma and "helpfully" warns that it did not work.

In crimson, the changes to store bench are following the recommended
action in the warning produced by clang.

Signed-off-by: Alex Ainscow <aainscow@uk.ibm.com>