mgr/cephadm: added variable check for bootstrap version
The mechanism relies on the variable bootstrap_versioned_stored in
the main CephadmOrchestrator class in Cephadm. This boolean variable
is initially false and is changed to true whenever a bootstrap version was
successfully added to the sqlite table. This mechanism was added to
allow quick check for if bootstrap version is added which will save on
overhead since functions that try to add the bootstrap version exist in
the Cephadm serve loop which is called every 60 seconds.

Signed-off-by: Timothy Q Nguyen <timqn22@gmail.com>