osd/scrub: fix blocked scrub accounting
The labeled perf counter <label>_chunk_busy is incremented whenever
the selected scrub chunk range is not available, and the scrubber
must wait for ongoing I/O to complete. This is as it should be.

<label>_locked_object should not be incremented in parallel,
serving no purpose. Instead - it should count the times the
'wait for chunk to be free' takes longer than the configured
grace period.

Fixes: https://tracker.ceph.com/issues/73415
Signed-off-by: Ronen Friedman <rfriedma@redhat.com>